Code Cleanup

+++ src/main/src/main/java/bletch/tektopiathief/entities/ai/EntityAIUseDoor.java
@@ -15,85 +15,85 @@ import net.tangotek.tektopia.entities.EntityVillageNavigator;
 import net.tangotek.tektopia.pathing.PathNavigateVillager2;
 
 public class EntityAIUseDoor extends EntityAIBase {
-	protected EntityLiving entity;
-	protected BlockDoor doorBlock;
-	protected BlockPos doorPosition;
-	int closeTimer;
+    protected EntityLiving entity;
+    protected BlockDoor doorBlock;
+    protected BlockPos doorPosition;
+    int closeTimer;
 
-	public EntityAIUseDoor(EntityLiving entitylivingIn) {
-		this.doorPosition = BlockPos.ORIGIN;
-		this.entity = entitylivingIn;
-	}
+    public EntityAIUseDoor(EntityLiving entitylivingIn) {
+        this.doorPosition = BlockPos.ORIGIN;
+        this.entity = entitylivingIn;
+    }
 
-	public boolean shouldExecute() {
-		if (!this.entity.collidedHorizontally) {
-			return false;
-		}
+    public boolean shouldExecute() {
+        if (!this.entity.collidedHorizontally) {
+            return false;
+        }
 
-		PathNavigateVillager2 pathNavigate = (PathNavigateVillager2)this.entity.getNavigator();
-		Path path = pathNavigate.getPath();
+        PathNavigateVillager2 pathNavigate = (PathNavigateVillager2) this.entity.getNavigator();
+        Path path = pathNavigate.getPath();
 
-		if (path != null && !path.isFinished() && pathNavigate.getEnterDoors()) {
-			for(int i = 0; i < Math.min(path.getCurrentPathIndex() + 2, path.getCurrentPathLength()); ++i) {
-				PathPoint pathpoint = path.getPathPointFromIndex(i);
-				this.doorPosition = new BlockPos(pathpoint.x, pathpoint.y + 1, pathpoint.z);
+        if (path != null && !path.isFinished() && pathNavigate.getEnterDoors()) {
+            for (int i = 0; i < Math.min(path.getCurrentPathIndex() + 2, path.getCurrentPathLength()); ++i) {
+                PathPoint pathpoint = path.getPathPointFromIndex(i);
+                this.doorPosition = new BlockPos(pathpoint.x, pathpoint.y + 1, pathpoint.z);
 
-				if (this.entity.getDistanceSq((double)this.doorPosition.getX(), this.entity.posY, (double)this.doorPosition.getZ()) <= 2.25D) {
-					this.doorBlock = this.getBlockDoor(this.doorPosition);
+                if (this.entity.getDistanceSq(this.doorPosition.getX(), this.entity.posY, this.doorPosition.getZ()) <= 2.25D) {
+                    this.doorBlock = this.getBlockDoor(this.doorPosition);
 
-					if (this.doorBlock != null) {
-						return true;
-					}
-				}
-			}
+                    if (this.doorBlock != null) {
+                        return true;
+                    }
+                }
+            }
 
-			this.doorPosition = (new BlockPos(this.entity)).up();
-			this.doorBlock = this.getBlockDoor(this.doorPosition);
-			return this.doorBlock != null;
-		}
+            this.doorPosition = (new BlockPos(this.entity)).up();
+            this.doorBlock = this.getBlockDoor(this.doorPosition);
+            return this.doorBlock != null;
+        }
 
-		return false;
-	}
+        return false;
+    }
 
-	public void startExecuting() {
-		LoggerUtils.info("EntityAIUseDoor - startExecuting called", true);
+    public void startExecuting() {
+        LoggerUtils.info("EntityAIUseDoor - startExecuting called", true);
 
-		this.closeTimer = 25;
-		this.toggleDoor(true);
-	}
+        this.closeTimer = 25;
+        this.toggleDoor(true);
+    }
 
-	public boolean shouldContinueExecuting() {
-		return this.closeTimer >= 0;
-	}
+    public boolean shouldContinueExecuting() {
+        return this.closeTimer >= 0;
+    }
 
-	public void updateTask() {
-		--this.closeTimer;
+    public void updateTask() {
+        --this.closeTimer;
 
-		if (this.closeTimer == 0 && !this.isDoorClear()) {
-			this.toggleDoor(true);
-			this.closeTimer = 25;
-		}
+        if (this.closeTimer == 0 && !this.isDoorClear()) {
+            this.toggleDoor(true);
+            this.closeTimer = 25;
+        }
 
-		super.updateTask();
-	}
+        super.updateTask();
+    }
 
-	public void resetTask() {
-		LoggerUtils.info("EntityAIUseDoor - resetTask called", true);
+    public void resetTask() {
+        LoggerUtils.info("EntityAIUseDoor - resetTask called", true);
 
-		this.toggleDoor(false);
-	}
+        this.toggleDoor(false);
+    }
 
-	protected BlockDoor getBlockDoor(BlockPos pos) {
-		IBlockState iblockstate = this.entity.world.getBlockState(pos);
-		Block block = iblockstate.getBlock();
-		return block instanceof BlockDoor && iblockstate.getMaterial() == Material.WOOD ? (BlockDoor)block : null;
-	}
+    protected BlockDoor getBlockDoor(BlockPos pos) {
+        IBlockState iblockstate = this.entity.world.getBlockState(pos);
+        Block block = iblockstate.getBlock();
+        return block instanceof BlockDoor && iblockstate.getMaterial() == Material.WOOD ? (BlockDoor) block : null;
+    }
 
-	protected boolean isDoorClear() {
-		return this.entity.world.getEntitiesWithinAABB(EntityVillageNavigator.class, new AxisAlignedBB(this.doorPosition)).isEmpty();
-	}
+    protected boolean isDoorClear() {
+        return this.entity.world.getEntitiesWithinAABB(EntityVillageNavigator.class, new AxisAlignedBB(this.doorPosition)).isEmpty();
+    }
 
-	protected void toggleDoor(boolean open) {
-		this.doorBlock.toggleDoor(this.entity.world, this.doorPosition, open);
-	}
+    protected void toggleDoor(boolean open) {
+        this.doorBlock.toggleDoor(this.entity.world, this.doorPosition, open);
+    }
 }
